+++ fixtures/coverall/src/lib.rs
@@ -358,4 +358,4 @@ impl ThreadsafeCounter {
     }
 }
 
-include!(concat!(env!("OUT_DIR"), "/coverall.uniffi.rs"));
+uniffi::include_scaffolding!("coverall");