+++ uniffi_bindgen/src/bindings/ruby/mod.rs
@@ -2,11 +2,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use std::{io::Write, process::Command};
+use std::process::Command;
 
 use anyhow::{Context, Result};
 use camino::Utf8Path;
-use fs_err::File;
+use fs_err as fs;
 
 pub mod gen_ruby;
 mod test;
@@ -24,9 +24,7 @@ pub fn write_bindings(
     try_format_code: bool,
 ) -> Result<()> {
     let rb_file = out_dir.join(format!("{}.rb", ci.namespace()));
-    let mut f = File::create(&rb_file)?;
-    write!(f, "{}", generate_ruby_bindings(config, ci)?)
-        .context("Failed to write generated Ruby bindings")?;
+    fs::write(&rb_file, generate_ruby_bindings(config, ci)?)?;
 
     if try_format_code {
         if let Err(e) = Command::new("rubocop").arg("-A").arg(&rb_file).output() {